gdb: fix broken prompt
The escape sequences in the gdb prompt confuse the internal readline
function making it think that the prompt is longer than what it is.
When using readline shortcuts like "jump to start of line" after having
wrote a somewhat long line of text with this broken config the cursor
gets incorrectly placed.

Use the `extended-prompt` config in place of `prompt` with \[ and \]
special characters to indicate to gdb what gets printed and what not and
solve the aforementioned problem.