gdb: add reference to gdb styling bug and implement workaround
If the LLVM prompt is invoked in the stop hook it will fill the list
cache with an unstyled version of the source.
The workaround for it is to flush the source cache at the end of every
hook-stop invocation.
Once the bug is solved on gdb's side it's better to let gdb use its
cache and remove the workaround.