vim: avoid interacting with server when invoking :SvnFiles
The `svn ls` command always contacts the SVN server to ask for the
command's output, but that is unneeded since all the info is already
contained in the local DB.
This is not ideal because the server may not be reachable at all times,
and crunching this data locally is anyway faster.

Change the command used to provide to FZF the file list so that it uses
the DB instead of the `svn ls` command.