w3m: make redirect_duckduckgo.cgi portable
POSIX awk does not allow using length() on an array to calculate its
number of cells, only gawk does that.
Since split() returns the number we care about use it in place of
invoking length(), making the script runnable on any version of awk
(tested with plan9's awk which should be old enough).